Real random number generator ready

+++ FlashCardsiOS/RandomNumberGenerator.h
@@ -10,6 +10,6 @@
 
 @protocol RandomNumberGenerator <NSObject>
 
--(float) next;
+-(int) next;
 
 @end
